% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{lout}
\alias{lout}
\title{Number of samples of the control parameter}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
lout
}
\description{
Called 'L' in the manuscript, specifies the number of evenly spaced samples from the control (bifurcation) parameter to be used.
}
\keyword{datasets}
