\name{lsp.model.example}
\alias{lsp.model.example}
\title{Load LSP model example code.}
\usage{
lsp.model.example(example.file)
}
\arguments{
  \item{example.file}{example file path relative to
  packaged root.}
}
\value{
text of model formulated in LSP language.
}
\description{
Load LSP model example code.
}
\examples{
lsp.model.example('extdata/knapsack.txt')
}

