% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving-functions.R
\docType{data}
\name{ODEs}
\alias{ODEs}
\alias{.doublewell}
\alias{doublewell}
\alias{.SIS}
\alias{SIS}
\alias{.genereg}
\alias{genereg}
\alias{.mutualistic}
\alias{mutualistic}
\title{Functions to standardize simulations of ODE/SDEs}
\format{
An object of class \code{list} of length 11.

An object of class \code{list} of length 7.

An object of class \code{list} of length 10.

An object of class \code{list} of length 16.
}
\usage{
.doublewell

doublewell(t, x, params)

.SIS

SIS(t, x, params)

.genereg

genereg(t, x, params)

.mutualistic

mutualistic(t, x, params)
}
\arguments{
\item{t}{An arbitrary time.}

\item{x}{The current state of the system.}

\item{params}{The list of model parameters.}
}
\value{
A vector of derivatives
}
\description{
Paired functions and lists of parameters to standardize using deSolve's ode() to compute ground truth simulations/solutions of several ODEs on various networks.
}
\details{
Names without dots are functions which compute derivatives and are in deSolve's standard form. Each function can be used in a one-time way to compute the derivative of a system given an arbitrary time t, current state x, and parameters. More commonly, passed to deSolve's ode() as the model to be simulated.
Dot names are lists of standard model parameters. Pass the required coupling strength like `c(.name, list(D = D))`. The adjacency matrix is also required for solutions on networks; pass it in the same way. If analyzing a single variable, x should have length 1 and pass A = matrix(0).
}
\examples{
library(parallel)
ncores <- detectCores()-1
library(igraph)
library(deSolve)
g <- readRDS("../data/dolphin.rds")
N <- vcount(g)
A <- as_adj(g, "both", sparse = FALSE)
k <- degree(g)
times <- 0:15
params <- c(.doublewell, list(A = A))
control <- list(times = times, ncores = ncores)
X <- solve_in_range(params$Ds, "D", doublewell, rep(params$xinit.low, N), params, control, "ode")
}
\keyword{datasets}
