% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving-functions.R
\name{preallocate_noise}
\alias{preallocate_noise}
\title{Preallocate noise for sde()}
\usage{
preallocate_noise(sigma, ntimesteps, nnodes)
}
\arguments{
\item{sigma}{The standard deviation of the noise process}

\item{ntimesteps}{The number of simulation time steps (probably the output of determine_ntimesteps())}

\item{nnodes}{The number of SDEs (i.e., number of nodes in a network)}
}
\value{
A matrix
}
\description{
Generates a ntimesteps x nnodes matrix of Gaussian random numbers to use for solving SDEs.
}
\examples{
ntimesteps <- determine_ntimesteps(times = 0:10, deltaT = 0.01)
W <- preallocate_noise(0.01, ntimesteps, 5)
}
